requirements.txt
streamlit
spacy
python-docx
reportlab
openai
google-generativeai
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz#egg=en_core_web_sm
